+++ NEWS.md
@@ -3,6 +3,7 @@
 Bug fixes:
 
 * Fixed bug in `FindMotifs()` when using only one region as input ([#732](https://github.com/timoast/signac/issues/732))
+* Add check for correct number of columns in fragment file ([#748](https://github.com/timoast/signac/issues/748))
 
 Other changes: